==> 3_summary_table_predictors.R <==
# 3_summary_table_predictors.R
# summary table of the predictors including levels and log-transform
# Jan 2021

# get the data
load('results/var_list_clintrials.RData') # from 2_elasticnet_model_samplesize_clintrial.R

# make table for latex


# could make into PDF Rmarkdown
