==> 98_summary_table_predictors.Rmd <==
---
title: "Supplement Table 1: Predictor variables in the two trial registries"
output: pdf_document
geometry: margin=2cm
---

```{r, include=FALSE}
knitr::opts_chunk$set(echo = FALSE, warning=FALSE, message=FALSE, error=FALSE, comment='', dpi=400)
options(width=1000) # Wide pages
options(scipen=999) # avoid scientific presentation

library(dplyr)
library(stringr)
library(flextable)

## get the data

# 0) labels
load('data/labels.RData') # for mandatory

# a) clintrials.gov
load('results/var_list_clintrials.RData') # from 2_elasticnet_model_samplesize_clintrial.R
# add mandatory, needs some management
vars = left_join(vars, mandatory_clintrials, by=c('variables'='var'))
clintrials_vars = select(vars, variables, categories, mandatory, log) # column ordering

# b) ANZCTR
load('results/var_list_anzctr.RData') # from 2_elasticnet_model_samplesize_ANZCTR.R
# add mandatory, needs some management
vars = left_join(vars, mandatory_anzctr, by=c('variables'='var'))
anzctr_vars = select(vars, variables, categories, mandatory, log) # column ordering
```


```{r, include=FALSE}
## add nicer variable names
load('data/labels.RData') # from '0_labels.R'
# add a few more labels
more_labels = read.table(sep=',', header=TRUE, stringsAsFactors = FALSE, text='
term,label
age_limit_max,Maximum age limit
age_limit_min,Minimum age limit
volunteers,Healthy volunteers
provisional,Provisional results
ccode1,Area
study_status,Study status
intervention_code,Intervention
study_design_time,Study design
study_design_observational,Study time
lead_sponsor_class,Lead sponsor
')
table_names_clintrials = bind_rows(table_names_clintrials, more_labels)
table_names_anzctr = bind_rows(table_names_anzctr, more_labels)

# merge with data
clintrials_vars = left_join(clintrials_vars, table_names_clintrials, by=c("variables" = "term")) %>%
  mutate(Variable = ifelse(is.na(label)==TRUE, str_to_sentence(variables), label)) %>% # sentence case if no label
  select(Variable, categories, mandatory, log) 
anzctr_vars = left_join(anzctr_vars, table_names_anzctr, by=c("variables" = "term")) %>%
  mutate(Variable = ifelse(is.na(label), str_to_sentence(variables), label)) %>%
  select(Variable, categories, mandatory, log) 

```

\thispagestyle{empty} 

The tables below show the predictor variables available for the two trial registries that were used in the multiple regression models. 

The "categories" column gives the categories for factors. 

The "mandatory" column indicates whether the variable has to be completed on the databases.

The "log" column indicates whether the variable was log-transformed (base 2). 

Study status was only used for the models of actual sample size and not the target sample size, this is because the study status was likely to change after the target sample size was set.

## clintrials.gov

```{r}
# add whether they are mandatory or optional?
ftab = flextable(clintrials_vars) %>%
  fontsize(part="all", size=10) %>%
  width(width=c(2.1, 4.3, 0.3, 0.3)) %>%
  theme_box()
ftab
```
\scriptsize

`+` = Mandatory after January 2017

\normalsize

\thispagestyle{empty} 

## ANZCTR

```{r}
ftab = flextable(anzctr_vars) %>%
  fontsize(size = 10, part = "all") %>%
  width(width=c(2, 5, 0.4, 0.3)) %>%
  theme_box()
ftab
```

\thispagestyle{empty}
